05a: Use long for IDs and use real data

+++ src/com/ruszki/aoc2025/e05/Ingredient.kt
@@ -1,3 +1,3 @@
 package com.ruszki.aoc2025.e05
 
-data class Ingredient(val id: UInt) {}
+data class Ingredient(val id: ULong) {}
