+++ src/com/ruszki/aoc2025/e05/IngredientIdRange.kt
@@ -1,6 +1,6 @@
 package com.ruszki.aoc2025.e05
 
-class IngredientIdRange(val start: UInt, val end: UInt) {
+class IngredientIdRange(val start: ULong, val end: ULong) {
     fun includes(ingredient: Ingredient): Boolean {
         return ingredient.id in start..end
     }
